# source-code/backend/docs/openapi.json
{
  "components": {
    "schemas": {
      "AccountResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "parent",
              "student"
            ],
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "role"
        ],
        "title": "AccountResponse",
        "type": "object"
      },
      "AccountSummaryResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "parent",
              "student"
            ],
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "role"
        ],
        "title": "AccountSummaryResponse",
        "type": "object"
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "newPassword": {
            "maxLength": 200,
            "minLength": 12,
            "title": "Newpassword",
            "type": "string"
          }
        },
        "required": [
          "newPassword"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
...
      "CreateAccountRequest": {
        "properties": {
          "grade": {
            "anyOf": [
              {
                "enum": [
                  "Grade 10",
                  "Grade 11"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "level": {
            "anyOf": [
              {
                "const": "iGCSE",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "parentId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentid"
          },
          "password": {
            "maxLength": 200,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "progression": {
            "items": {
              "type": "string"
            },
            "maxItems": 6,
            "title": "Progression",
            "type": "array"
          },
          "role": {
            "enum": [
              "parent",
              "student"
            ],
            "title": "Role",
            "type": "string"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Subjects",
            "type": "array"
          },
          "term": {
            "anyOf": [
              {
                "enum": [
                  "Term1",
                  "Term2",
                  "Term3"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          },
          "username": {
            "maxLength": 40,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "name",
          "password",
          "role"
        ],
        "title": "CreateAccountRequest",
        "type": "object"
      },
      "ErrorItem": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorItem",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorItem"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "csrfToken": {
            "title": "Csrftoken",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user",
          "csrfToken"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "PortalStateResponse": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountSummaryResponse"
            },
            "title": "Accounts",
            "type": "array"
          },
          "assignments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Assignments",
            "type": "array"
          },
          "attempts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Attempts",
            "type": "array"
          },
          "catalog": {
            "$ref": "#/components/schemas/PortalCatalogResponse"
          },
          "coverage": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Coverage",
            "type": "array"
          },
          "documents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Documents",
            "type": "array"
          },
          "drafts": {
            "additionalProperties": true,
            "title": "Drafts",
            "type": "object"
          },
          "exams": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Exams",
            "type": "array"
          },
          "plans": {
            "additionalProperties": true,
            "title": "Plans",
            "type": "object"
          },
          "questionBank": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Questionbank",
            "type": "array"
          },
          "questions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Questions",
            "type": "array"
          },
          "reviews": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Reviews",
            "type": "array"
          },
          "students": {
            "items": {
              "$ref": "#/components/schemas/StudentResponse"
            },
            "title": "Students",
            "type": "array"
          },
          "subjects": {
            "items": {
              "$ref": "#/components/schemas/SubjectPresentationResponse"
            },
            "title": "Subjects",
            "type": "array"
          },
          "units": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Units",
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user",
          "catalog",
          "accounts",
          "subjects",
          "students"
        ],
        "title": "PortalStateResponse",
        "type": "object"
      },
...
      "ProgressionResponse": {
        "properties": {
          "grade": {
            "title": "Grade",
            "type": "string"
          },
          "term": {
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "grade",
          "term"
        ],
        "title": "ProgressionResponse",
        "type": "object"
      },
      "StudentResponse": {
        "properties": {
          "courses": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SubjectCourseResponse"
            },
            "title": "Courses",
            "type": "object"
          },
          "grade": {
            "title": "Grade",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "initial": {
            "title": "Initial",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "needsConfiguration": {
            "title": "Needsconfiguration",
            "type": "boolean"
          },
          "parentId": {
            "title": "Parentid",
            "type": "string"
          },
          "progression": {
            "items": {
              "$ref": "#/components/schemas/ProgressionResponse"
            },
            "title": "Progression",
            "type": "array"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "title": "Subjects",
            "type": "array"
          },
          "term": {
            "title": "Term",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "initial",
          "parentId",
          "level",
          "grade",
          "term",
          "progression",
          "subjects",
          "courses",
          "needsConfiguration"
        ],
        "title": "StudentResponse",
        "type": "object"
      },
      "SubjectCourseResponse": {
        "properties": {
          "level": {
            "title": "Level",
            "type": "string"
          },
          "syllabus": {
            "title": "Syllabus",
            "type": "string"
          }
        },
        "required": [
          "level",
          "syllabus"
        ],
        "title": "SubjectCourseResponse",
        "type": "object"
      },
...
      "UiFeaturesResponse": {
        "properties": {
          "allowed": {
            "const": true,
            "title": "Allowed",
            "type": "boolean"
          },
          "user": {
            "$ref": "#/components/schemas/UiFeaturesUserResponse"
          }
        },
        "required": [
          "allowed",
          "user"
        ],
        "title": "UiFeaturesResponse",
        "type": "object"
      },
      "UiFeaturesUserResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "const": "admin",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "name",
          "role"
        ],
        "title": "UiFeaturesUserResponse",
        "type": "object"
      },
      "UpdateStudentRequest": {
        "properties": {
          "grade": {
            "enum": [
              "Grade 10",
              "Grade 11"
            ],
            "title": "Grade",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "level": {
            "const": "iGCSE",
            "title": "Level",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "parentId": {
            "format": "uuid",
            "title": "Parentid",
            "type": "string"
          },
          "progression": {
            "items": {
              "type": "string"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Progression",
            "type": "array"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Subjects",
            "type": "array"
          },
          "term": {
            "enum": [
              "Term1",
              "Term2",
              "Term3"
            ],
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "parentId",
          "level",
          "grade",
          "term",
          "progression",
          "subjects"
        ],
        "title": "UpdateStudentRequest",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mustChangePassword": {
            "title": "Mustchangepassword",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "parent",
              "student"
            ],
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "role",
          "mustChangePassword"
        ],
        "title": "UserResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AKURU API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/accounts": {
      "post": {
        "operationId": "create_account_api_v1_admin_accounts_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "akuru_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Akuru Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Create Account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/admin/students": {
      "post": {
        "operationId": "update_student_api_v1_admin_students_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "akuru_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Akuru Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Update Student",
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/admin/ui-features": {
      "get": {
        "operationId": "ui_features_access_api_v1_admin_ui_features_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "akuru_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Akuru Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UiFeaturesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Ui Features Access",
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "change_password_api_v1_auth_change_password_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "akuru_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Akuru Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Change Password",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Login",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "akuru_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Akuru Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "me_api_v1_auth_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "akuru_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Akuru Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "Me",
        "tags": [
          "authentication"
        ]
      }
    },
...
    "/api/v1/state": {
      "get": {
        "operationId": "state_api_v1_state_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "akuru_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Akuru Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          }
        },
        "summary": "State",
        "tags": [
          "accounts"
        ]
      }
    },
...
  }
}